FIx: Se realizaron cambios y arreglos: carrito mantiene contador entre páginas y se paso  el carrito en un archivo aparte: carrito.js
