+++ contacto.html
@@ -26,24 +26,26 @@
         <nav class="navbar">
           <ul class="navbar__list">
             <li class="navbar__item">
-
               <a href="./index.html" class="navbar__link">Inicio</a>
-
             </li>
             <li class="navbar__item">
               <a href="./productos.html" class="navbar__link">Productos</a>
             </li>
             <li class="navbar__item">
-              <a href="./contacto.html" class="navbar__link--active">Contacto</a>
+              <a href="./contacto.html" class="navbar__link--active"
+                >Contacto</a
+              >
             </li>
 
-             <li class="navbar__item">
-              <a href="#" class="navbar__link"> <i class="fa-solid fa-cart-shopping"></i></a>
+            <li class="navbar__item">
+              <a href="#" class="navbar__link">
+                <i class="fa-solid fa-cart-shopping"></i
+              ></a>
               <span class="header__carrito-contador">0</span>
             </li>
 
-             <li class="navbar__item">
-              <a href="#" class="navbar__link"> 
+            <li class="navbar__item">
+              <a href="#" class="navbar__link">
                 <i class="fa-solid fa-magnifying-glass"></i>
               </a>
             </li>
@@ -85,7 +87,13 @@
           <div class="contact-form__group">
             <label for="mensaje" class="contact-form__label">Mensaje</label>
             <textarea
-              name="mensaje" id="mensaje" cols="30" rows="5" placeholder="Mensaje..." class="contact-form__textarea"></textarea>
+              name="mensaje"
+              id="mensaje"
+              cols="30"
+              rows="5"
+              placeholder="Mensaje..."
+              class="contact-form__textarea"
+            ></textarea>
             <p></p>
           </div>
 
@@ -178,10 +186,10 @@
               <a class="footer__link" href="#">ventas@hermanosjota.com.ar</a>
             </li>
             <li class="footer__item">
-              <a class="footer__link" href="#">Av. San Juan 2847
-C1232AAB — Barrio de San Cristóbal
-Ciudad Autónoma de Buenos Aires
-Argentina</a>
+              <a class="footer__link" href="#"
+                >Av. San Juan 2847 C1232AAB — Barrio de San Cristóbal Ciudad
+                Autónoma de Buenos Aires Argentina</a
+              >
             </li>
           </ul>
         </div>
@@ -210,6 +218,6 @@ Argentina</a>
         <p>© 2025 Hermanos Jota. Todos los derechos reservados.</p>
       </div>
     </footer>
-        <script src="/js/data.js"></script>
+    <script type="module" src="./js/app.js"></script>
   </body>
 </html>
