+++ index.html
@@ -197,10 +197,10 @@
               <a class="footer__link" href="#">ventas@hermanosjota.com.ar</a>
             </li>
             <li class="footer__item">
-              <a class="footer__link" href="#">Av. San Juan 2847
-C1232AAB — Barrio de San Cristóbal
-Ciudad Autónoma de Buenos Aires
-Argentina</a>
+              <a class="footer__link" href="#"
+                >Av. San Juan 2847 C1232AAB — Barrio de San Cristóbal Ciudad
+                Autónoma de Buenos Aires Argentina</a
+              >
             </li>
           </ul>
         </div>
@@ -229,6 +229,6 @@ Argentina</a>
         <p>© 2025 Hermanos Jota. Todos los derechos reservados.</p>
       </div>
     </footer>
-    <script src="/js/data.js"></script>
+    <script type="module" src="./js/app.js"></script>
   </body>
 </html>
