+++ producto.html
@@ -203,6 +203,6 @@ Argentina</a>
         <p>© 2025 Hermanos Jota. Todos los derechos reservados.</p>
       </div>
     </footer>
-        <script src="/js/data.js"></script>
+        <script type="module" src="./js/app.js"></script>
   </body>
 </html>
